requirements.txt
aiohttp==3.8.4
python-telegram-bot==20.3
Flask==2.2.3
Werkzeug==2.2.3
APScheduler==3.9.1.post1
SQLAlchemy==2.0.12
python-dotenv==1.0.0
psycopg2-binary
imapclient==2.3.0
email-validator==1.3.1